backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
requests
beautifulsoup4